+++ EveLib.EveCrest/Models/IInferrableId.cs
@@ -7,6 +7,6 @@
         ///     Gets or sets the inferred identifier.
         /// </summary>
         /// <value>The inferred identifier.</value>
-        int InferredId { get; set; }
+        long InferredId { get; set; }
     }
 }
